Upgrade to lib-external-provisioner v8.
This fixes local-path-provisioner not properly shutting down on
sigterm.

In v4, ProvisionController.Run() ignored the channel given to it,
thus would never stop.  This is addressed in v8 via context being
threaded down through the relevant k8s api interactions.

As to what v8 brings; various performance fixes and improvements.
For this use case, the only relevant thing is the proper shutdown.