fix: use kong for CLI parsing to unify and simplify parsing.
The previous usage of cli lead to validation, typing, and flag definitions being
all over the code.  Via kong, we can centralize the parsing restrictions into
one spot, simplifying the code that consumes the CLI.

As to the relevance of this: this is prep work to bring the CLI more under control
(complexity wise) so new features can be added.