Fix: depriv the helper by default
In b44b2c3b5d, support for setting the service account was added;
it defaulted to using the service account that the provisioner
itself used.  IE, the helperPod could mutate K8 objects, launch
it's own pods, etc.  This is not needed as a default.

Being that the end user can specify the pod.Spec template themselves
(thus allowing them to set serviceAccountName themselves), the CLI- and this
provisioner- don't need to be involved, or try to override it.

helperPod's new default is to use 'default' serviceAccount of the
target namespace and to set `automountServiceAccountToken: false`.

If you've got a helperPod setup that needs to query or mutate K8s, just
set whatever serviceAccount you need in the helperPod's spec and
automountServiceAccountToken will be left as true.